# eric-am-onboarding-server-api/src/main/resources/private/api.yaml
...
openapi: 3.0.3
info:
  version: '1.0'
  title: Application Manager Onboarding Service
servers:
  - url: '{protocol}://{hostname}/{basePath}'
    variables:
      protocol:
        default: https
      hostname:
        default: ${INGRESS_HOST_ADDRESS}/vnfm/onboarding
      basePath:
        default: api/v1
...
paths:
...
  /internal/packages:
    post:
      tags:
       - Internal Package Controller
      description: Retrieve a resource with id and returns a VnfResource
      parameters:
        - name: helmChartURI
          in: query
          required: true
          schema:
            type: string
      operationId: v1PackagesPost
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                type: string
  /packages:
    get:
      tags:
        - Packages
      summary: V1Packages_GET
      parameters:
        - name: filter
          description: |
            Attribute-based filtering. The Onboarding shall support receiving this parameter
            as part of the URI query string. The VNFM/OSS may supply this parameter. All
            attribute names that appear in the AppPackageQuery and in data types referenced
            from it shall be supported by the Onboarding in the filter expression
          in: query
          required: false
          schema:
            type: string
        - name: nextpage_opaque_marker
          description: |
            Marker to obtain the page number of a paged response.
          in: query
          required: false
          schema:
            type: string
        - name: size
          description: |
            The number of items per page. The default value is 15. Should be a positive number no greater than 100.
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                title: AppPackageList
                properties:
                  packages:
                    type: array
                    items:
                      $ref: '#/components/schemas/AppPackageResponse'
          headers:
            Link:
              description: >
                Reference to other resources. Used for paging in the present
                document.
              schema:
                type: string
                maximum: 1
                minimum: 0
            PaginationInfo:
              description: >
                Page information of result set, includes
                  number - the returned page number
                  size - the number of results returned in the page
                  totalPages - the total number of pages for the query
                  totalElements - the total number of elements for the query
              schema:
                type: string
                maximum: 1
                minimum: 0
  '/packages/{id}':
    get:
      tags:
        - Packages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 'The Id of the package'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppPackageResponse'
        '400':
          description: BAD REQUEST
        '404':
          description: PACKAGE NOT FOUND
        '500':
          description: INTERNAL SERVER ERROR
    delete:
      tags:
        - Packages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ''
      responses:
        '204':
          description: SUCCESSFULLY DELETED RESOURCE
        '400':
          description: BAD REQUEST
        '404':
          description: PACKAGE NOT FOUND
        '500':
          description: INTERNAL SERVER ERROR
...
components:
  schemas:
    OperationDetailResponse:
      description: Operation Detail response for package
      type: object
      title: OperationDetailResponse
      properties:
        operationName:
          type: string
        supported:
          type: boolean
        error:
          type: string
...
    HelmPackage:
      type: object
      properties:
        chartUrl:
          description: |
            The URL to the chart.
          type: string
        chartName:
          description: |
            The name of the chart.
          type: string
        chartVersion:
          description: |
            The version of the chart.
          type: string
        chartType:
          description: |
            The type of the chart.
          type: string
          enum:
            - CRD
            - CNF
        priority:
          description: |
            The order in which the charts will be installed in order of ascendency.
          type: integer
        chartArtifactKey:
          description: |
            The artifact key of the chart.
          type: string
...
    AppPackageResponse:
      properties:
        appPkgId:
          description: |
            Id of the app package.
          type: string
        appDescriptorId:
          description: >
            The app descriptor id
          type: string
        appProvider:
          description: >
            The app provider
          type: string
        appProductName:
          description: |
            The application product name.
          type: string
        appSoftwareVersion:
          description: >
            The application software version
          type: string
        descriptorVersion:
          description: >
            The descriptor version
          type: string
        descriptorModel:
          description: |
            The descriptor model
          type: object
        helmPackageUrls:
          description: |
            The list of helm package urls
          type: array
          items:
            $ref: '#/components/schemas/HelmPackage'
        imagesURL:
          description: |
            The URL to the images.
          type: array
          items:
            type: object
        packageFiles:
          description: |
            List of package files
          type: array
          items:
            type: object
        onboardingState:
          description: >
            The onboarding state.
          type: string
        usageState:
          description: |
            The Usage state for the package.
          type: string
          enum:
            - NOT_IN_USE
            - IN_USE
        packageSecurityOption:
          type: string
          description: |
            The signature type of the package.
          enum:
            - OPTION_1
            - OPTION_2
            - UNSIGNED
        serviceModelId:
          description: >
            Contains the Id of the service model if package has been uploaded to Tosca-O service.
          type: string
        supportedOperations:
          type: array
          items:
            $ref: '#/components/schemas/OperationDetailResponse'
        problemDetails:
          description: >
            Problem details during onboarding
          type: string
...